Separating rules from the main PeepholeMutator impl
